+++ pkg/auth/web/user.go
@@ -35,22 +35,13 @@ func (ep *Endpoint) InitCreateUser(w http.ResponseWriter, r *http.Request) {
 	// It avoids filling in the form again after submissions errors.
 	userForm := ep.RestoreUserForm(r, web.CreateUserStoreKey)
 
-	// Restore previous flash data if exists
-	pf := ep.RestoreFlash(r)
-	if len(pf) > 0 {
-		ep.Log().Debug("Flash data", "current", spew.Sdump(pf))
-	}
-
 	// Req & Res
 	res := &tp.CreateUserRes{}
 	res.FromTransport(&userForm, "", nil)
 	res.Action = ep.userCreateAction()
 
-	// TEMP: Just for use test
-	pf = pf.AddItem(ep.MakeFlashItem("Appended to previous", web.InfoMT))
-
 	// Wrap response
-	wr := ep.OKRes(r, res)
+	wr := ep.OKRes(r, res, "[sample-msg] Init user creation")
 
 	// Template
 	ts, err := ep.TemplateFor(userRes, web.CreateTmpl)
@@ -104,9 +95,9 @@ func (ep *Endpoint) CreateUser(w http.ResponseWriter, r *http.Request) {
 	ep.ClearUserForm(r, w, web.CreateUserStoreKey)
 
 	// Flash message (sample test)
-	ep.StoreFlash(r, w, "Sample message", web.InfoMT)
+	ep.StoreFlash(r, w, "[sample-msg] User created", web.InfoMT)
 
-	ep.Redirect(w, r, "/")
+	ep.Redirect(w, r, UserPath())
 }
 
 // GetUsers web endpoint.
@@ -129,7 +120,7 @@ func (ep *Endpoint) GetUsers(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Wrap response
-	wr := ep.OKRes(r, res)
+	wr := ep.OKRes(r, res, "")
 
 	// Template
 	ts, err := ep.TemplateFor(userRes, web.IndexTmpl)
@@ -183,48 +174,6 @@ func (ep *Endpoint) localizeMessageID(l *i18n.Localizer, messageID string) (stri
 	return l.Localize(&i18n.LocalizeConfig{
 		MessageID: messageID,
 	})
-}
-
-// Flash
-func (ep *Endpoint) StoreFlash(r *http.Request, w http.ResponseWriter, message string, mt web.MsgType) (ok bool) {
-	s := ep.GetSession(r)
-
-	// Append to current ones
-	f := ep.RestoreFlash(r)
-	f = append(f, ep.MakeFlashItem(message, mt))
-
-	s.Values[web.FlashStoreKey] = f
-	err := s.Save(r, w)
-	if err != nil {
-		ep.Log().Error(err)
-		return true
-	}
-
-	return false
-}
-
-func (ep *Endpoint) RestoreFlash(r *http.Request) web.FlashSet {
-	s := ep.GetSession(r)
-	v := s.Values[web.FlashStoreKey]
-
-	user, ok := v.(web.FlashSet)
-	if ok {
-		ep.Log().Debug("Stored flash", "value", spew.Sdump(user))
-		return web.MakeFlashSet()
-	}
-
-	ep.Log().Info("No stored flash", "key", web.FlashStoreKey)
-	return web.MakeFlashSet()
-}
-
-func (ep *Endpoint) ClearFlash(r *http.Request, w http.ResponseWriter, message string, mt web.MsgType) (ok bool) {
-	s := ep.GetSession(r)
-	delete(s.Values, web.FlashStoreKey)
-	err := s.Save(r, w)
-	if err != nil {
-		return true
-	}
-	return false
 }
 
 // Form data session helpers